Add includePending option to getItem() method
This allows getItem() to be used in two ways:
1. Only fetch confirmed items.
2. Fetch both confirmed & pending items.

This fixes an issue where the Forge couldn't find an item to smelt via the HTTP API, this was noticed when a KOTA player tried to smelt an item, and the game's Forge integration went unresponsive, with the item never smelting.

Also fixed some variable name re-use, and the Forge will properly catch + report RPC errors during crafting.