Add maxInvalidScore option to config file
The user can now customize the maxInvalidScore, if they are encountering too many validation inconsistencies or false-positives, they can raise this score to help prevent issues, at the cost of allowing potentially malicious items to be present in the node for longer before being removed.

This also changes the default to 50 from 25, to prevent blockbook API slowness from causing inconsistencies, in addition to doubling the janitor interval time.