Add node connectivity awareness
Nodes will now intelligently categorize/disconnect peers, instead of just plain disconnecting them if they don't respond, a node can tell if a peer is online even if they are not reachable via received ping time, the majority of non-portforwarded user nodes are not reachable, so this will prevent disconnecting users and breaking handshakes.

This allows portforwarded nodes to mark non-reachable as "sendOnly", waiting for a ping to be received from them, instead of repeatedly trying to ping data to them.