Add off-by-default 'Full Node' and overall optimizations
This commit adds a new config option, which is disabled by default if the option is missing from the config.
The new option, "full node" allows the node operator to toggle if they perform full validations on the entire network; This is useful to enable for servers that query the 'full' network, such as Game servers, marketplaces for ZFIs, etc.

For regular users, this functionality is disabled, instead the Forge will solely validate items related to the operator; e.g: Items they own and/or created then transferred.

Alongside this massive performance & scalability boost for regular users, a large amount of under-the-hood optimizations and fixes have been done, greatly reducing the amount of processing load on the attached ZENZO Core client; GUI 'hiccups' have been greatly reduced, and daemon unresponsiveness fixed entirely, even on very large wallets.