Add "Lock Status" to the Wallet tab
This commit introduces a button which also acts as a status bar, it displays the current lock status of the Core wallet (Locked/Unlocked) if it detects that the Core wallet uses encryption.

If the Forge doesn't detect any encryption, it will hide the status bar, as it's not useful information.

If the wallet is encrypted but Unlocked, the Lock Status button can be clicked which will lock the Core wallet instantly, however, the Core wallet cannot be unlocked via this button, only locked.