Implement smart Operator Address handling
This commit redesigns the Operator Address handling to have a smart and superior "selection" and recovery capability;
The Forge can now handle many different startup and environment cases, instead of freezing up and hitting issues when an Address is incorrectly managed, or a wallet.dat change occurs.

- New installs will automatically create a new "Forge" address.

- Wallets with a previous Forge address, but a corrupted config, will make the Forge perform a "recover and select" case in which the old address is recovered, loaded and saved back to the config file.

- Wallets with multiple Forge addresses will be assessed by "priority", Forge addresses with non-zero account balances are prioritized to ensure the Forge only selects addresses that have previously been interacted with; to prevent address "duplicates" from causing harm.

In addition, the startup and initialization process is now faster and smoother for old and new users.