Add type-safety to the ZVM interpreter
Previously all input was considered under a single, universal "native_input" OPcode, this meant that it was very easy for HEX-encoded strings and Numericals to get "mixed up" if one looked like the other.

Now, ZFI-templated scripts are able to utilize type-safety to ensure a script's input will always be interpreted as intended by the developer, and developers may now mark HEX strings with "HEX:" to explicitly ensure the ZVM interprets the input as a string, this can be done automatically with the included util.js library file.