[ZFI Revamp] Step 2: Rewrite ZFI databasing
ZFIs are now written and read against the DB according to the new class structure, the 'pending' and 'unsigned' DBs have been removed in favor of the singular 'items' DB in-line with the backend revamp.

The Forge is NOT functional with this commit alone, so please, do not attempt to run and compile based on this commit.