[ZFI Revamp] Step 6: Introduce ZFI Versions, revamp approval logic, smarter pointer usage
Each ZFI now has a version, or a 'nonce' attached, which can tell you how many times it's been edited in the past, this allows nodes to know straight away if an item's copy is older or newer than a comparison of itself; making syncing extremely efficient by eliminating older versions from the network. This, for example, massively simplifies 'transfers' and the ability to edit items on-the-go.