[Core] [Init] Delete Forge repo & quit launch
The entirety of the Forge repo has been deleted and force-pushed to Github, and I have resigned from my position as Lead Developer.

April Fools!

This commit fixes a simple issue on slow / large wallets, which get stuck on the locking phase of the bootup, the frontend recognizes the slowness as a lock of wallet connection, and kicks the bucket with the "Setup" screen.
This should be remedied now!